% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubert.R
\name{cubert}
\alias{cubert}
\title{Cube Root a vector}
\usage{
cubert(x)
}
\arguments{
\item{x}{The vector to be cube rooted.}
}
\value{
A vector that is the square of \code{x}.
}
\description{
That's it -- this function just cube roots a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{cubert} function is quite simple.
     \item There's nothing else to say about \code{cubert}.
}
}
\examples{
cubert(1:10)
cubert(9)
}
