% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.R
\name{inverse}
\alias{inverse}
\title{Inverse a vector}
\usage{
inverse(x)
}
\arguments{
\item{x}{The vector to be inversed.}
}
\value{
A vector that is the inverse of \code{x}.
}
\description{
That's it -- this function just inverses a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{inverse} function is quite simple.
     \item There's nothing else to say about \code{inverse}.
}
}
\examples{
inverse(1:10)
inverse(9)
}
