% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powersexpanded-package.R
\docType{package}
\name{powersexpanded}
\alias{powersexpanded}
\alias{powersexpanded-package}
\title{Expanded Computation of Powers.}
\description{
This R package is an extended version of Vincenzo's class example. It provides functions to compute powers.
}
\note{
This package was created for learning purposes as part of STAT 547 hw09.
}
\author{
Victoria Michalowski
}
