% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squarert.R
\name{squarert}
\alias{squarert}
\title{Square Root a vector}
\usage{
squarert(x)
}
\arguments{
\item{x}{The vector to be square rooted.}
}
\value{
A vector that is the square root of \code{x}.
}
\description{
That's it -- this function just square roots a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{squarert} function is quite simple.
     \item There's nothing else to say about \code{squarert}.
}
}
\examples{
squarert(1:10)
squarert(16)
}
